Move RevisionDetector tests to new class (RevisionDetectorTest).
